Move low level lock/condvar/semaphore form libwasm_work to libc.
I've moved the declarations of these functions to a new header
(`threading_primitives.h`) which is then included in both
`wasm_workers.h` and in `threading.h`.

Add a some basic testing that these function are available under
pthreads as well as wasm workers.
